Removed the comment functions from the SqlLiteHelper, Added a CommentLoaderHelper to work with the CommentActivity
