+++ app/src/main/java/com/alfa/alfanieuws/Services/SqlLiteHelper.java
@@ -67,9 +67,7 @@ public class SqlLiteHelper extends SQLiteOpenHelper {
             try {
                 SQLiteDatabase db = this.getWritableDatabase();
                 Cursor cursor = db.rawQuery(query, null);
-                cursor.moveToFirst();
                 db.close();
-                cursor.close();
                 return cursor;
             } catch (Exception e) {
                 System.out.println("Something went wrong using the executeSQL functionality");
@@ -123,7 +121,7 @@ public class SqlLiteHelper extends SQLiteOpenHelper {
     // delete with where string in query
     // DB.delete(tafel, WHERE?);
 
-    private void delete(String tableName, String where) {
+    public void delete(String tableName, String where) {
         if(tableName != null && where != null){
             try {
                 SQLiteDatabase db = this.getWritableDatabase();
@@ -135,27 +133,5 @@ public class SqlLiteHelper extends SQLiteOpenHelper {
         }
         System.out.println("I need you to enter a table name as STRING and the where clausule as STRING");
 
-    }
-
-    //Comments
-    public void addComment(String name, String post_id, String comment){
-        SQLiteDatabase db = this.getWritableDatabase();
-        ContentValues cv = new ContentValues();
-
-        cv.put(RESPONSE_COLUMN_NAME, name);
-        cv.put(NEWS_COLUMN__ID, post_id);
-        cv.put(RESPONSE_COLUMN_TEXT, comment);
-        long result = db.insert(RESPONSE_TABLE_NAME,null, cv);
-        if(result == -1){
-            System.out.println("Failed to add Comment");
-        }else {
-            System.out.println("Comment added");
-        }
-    }
-
-    public Cursor getAllComments(String post_id){
-        SQLiteDatabase db = this.getWritableDatabase();
-        Cursor res = db.rawQuery("SELECT * FROM " + RESPONSE_TABLE_NAME + " WHERE _newsId = " + post_id, null);
-        return res;
     }
 }
